+++ provenance_lib/replay.py
@@ -13,7 +13,7 @@ from ._archive_parser import ProvNode
 from .parse import ProvDAG, UUID
 from ._usage_drivers import DRIVER_CHOICES, SUPPORTED_USAGE_DRIVERS, Usage
 from .util import FileName, camel_to_snake
-from .yaml_constructors import MetadataInfo
+from ._yaml_constructors import MetadataInfo
 
 from qiime2.sdk import PluginManager
 from qiime2.sdk.usage import UsageVariable
